requirements.txt
Django
requests
django-haystack
jieba
Whoosh